Add env variable to the customizable_forms plugin for adding additional custom forms that are outside the plugin and change the Slate Project plugin to use it
