+++ coldfront/config/plugins/customizable_forms.py
@@ -5,3 +5,5 @@ from coldfront.config.env import ENV
 INSTALLED_APPS += [
     'coldfront.plugins.customizable_forms',
 ]
+
+ADDITIONAL_CUSTOM_FORMS = ENV.list('ADDITIONAL_CUSTOM_FORMS', default=[])
