+++ coldfront/plugins/customizable_forms/custom/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from coldfront.plugins.customizable_forms.custom.views import ComputeView, PositConnectView, SlateProjectView, GeodeProjectView
+from coldfront.plugins.customizable_forms.custom.views import ComputeView, PositConnectView, GeodeProjectView
 
 
 urlpatterns = [
@@ -17,11 +17,6 @@ urlpatterns = [
         '<int:project_pk>/create/<int:resource_pk>/positconnect',
         PositConnectView.as_view(),
         name='posit-form'
-    ),
-    path(
-        '<int:project_pk>/create/<int:resource_pk>/slateproject',
-        SlateProjectView.as_view(),
-        name='slateproject-form'
     ),
     path(
         '<int:project_pk>/create/<int:resource_pk>/geode-projects',
