+++ coldfront/plugins/customizable_forms/validators.py
@@ -1,9 +1,6 @@
-import re
-import os
 import logging
 from django.core.exceptions import ValidationError
 from coldfront.plugins.ldap_user_info.utils import get_user_info
-from coldfront.core.allocation.models import AllocationAttribute
 from coldfront.core.utils.common import import_from_settings
 
 logger = logging.getLogger(__name__)
@@ -35,12 +32,6 @@ class ValidateAccountNumber():
             raise ValidationError(f'Format is not correct', code='invalid')
 
 
-class ValidateDirectoryName():
-    def __call__(self, value):
-        if re.search('^[0-9a-zA-Z_-]+$', value) is None:
-            raise ValidationError(f'Contains invalid character(s)', code='invalid')
-
-
 class ValidateUsername():
     def __call__(self, value):
         if not value:
@@ -49,26 +40,3 @@ class ValidateUsername():
         attribute = get_user_info(value, ['cn'])
         if not attribute.get('cn') or not attribute.get('cn')[0]:
             raise ValidationError('This username does not exist')
-
-
-class ValidateDupDirectoryName():
-    def __call__(self, value):
-        if not value:
-            return
-
-        directory_value = '/N/project/' + value
-        directory_names = AllocationAttribute.objects.filter(
-            allocation_attribute_type__name='Slate Project Directory'
-        ).values_list('value', flat=True)
-        for directory_name in directory_names:
-            if directory_name == directory_value:
-                raise ValidationError('This Slate Project directory name already exists')
-
-        if not os.path.isfile(os.path.join(SLATE_PROJECT_INCOMING_DIR, 'slate_projects.txt')):
-            logger.warning('slate_project.txt is missing. Skipping additional directory name checking')
-            return
-
-        with open(os.path.join(SLATE_PROJECT_INCOMING_DIR, 'slate_projects.txt'), 'r') as slate_projects:
-            for line in slate_projects:
-                if line.strip('\n') == value:
-                    raise ValidationError('This Slate Project directory name already exists')
