+++ coldfront/plugins/slate_project/templates/slate_project/slateproject.html
