+++ coldfront/plugins/slate_project/views.py
@@ -1,4 +1,6 @@
 import logging
+import requests
+from urllib import parse
 from django.urls import reverse
 from django.views.generic import ListView, View
 from django.contrib.auth.mixins import LoginRequiredMixin
@@ -7,14 +9,18 @@ from django.http import HttpResponseForbidden, HttpResponseRedirect
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render
 
-from coldfront.core.allocation.models import Allocation, AllocationAttribute
+from coldfront.core.allocation.models import Allocation, AllocationAttribute, AllocationAttributeType
 from coldfront.core.utils.mail import build_link
 from coldfront.plugins.slate_project import utils
 from coldfront.plugins.slate_project.forms import SlateProjectSearchForm
 from coldfront.core.project.models import Project
-from coldfront.core.utils.common import get_domain_url, import_from_settings
+from coldfront.core.utils.common import import_from_settings
 from coldfront.core.utils.mail import send_email_template
+from coldfront.plugins.slate_project.forms import SlateProjectForm
 
+
+SLATE_PROJECT_ALLOCATED_QUANTITY_THRESHOLD = import_from_settings('SLATE_PROJECT_ALLOCATED_QUANTITY_THRESHOLD', 120)
+SLATE_PROJECT_MOU_SERVER = import_from_settings('SLATE_PROJECT_MOU_SERVER', '')
 EMAIL_ENABLED = import_from_settings('EMAIL_ENABLED', False)
 if EMAIL_ENABLED:
     SLATE_PROJECT_EMAIL = import_from_settings('SLATE_PROJECT_EMAIL', '')
@@ -121,3 +127,79 @@ class RequestAccessEmailView(LoginRequiredMixin, View):
             return HttpResponseForbidden(reverse('project-list'))
 
         return HttpResponseRedirect(reverse('project-list'))
+
+
+class SlateProjectView:
+    form_class = SlateProjectForm
+    template_name = 'slate_project/slateproject.html'
+
+    def get_context_data(self, **kwargs):
+        context =  super().get_context_data(**kwargs)
+
+        project_obj = get_object_or_404(Project, pk=self.kwargs.get('project_pk'))
+        pi_username = project_obj.pi.username
+        context['total_pi_allocated_quantity'] = utils.get_pi_total_allocated_quantity(pi_username)
+        context['pi_allocated_quantity_threshold'] = SLATE_PROJECT_ALLOCATED_QUANTITY_THRESHOLD
+
+        return context
+
+    def form_valid(self, form):
+        form_data = form.cleaned_data
+
+        start_date = form_data.get('start_date', '')
+        start_date = start_date.strftime('%m/%d/%Y')
+
+        project_obj = get_object_or_404(Project, pk=self.kwargs.get('project_pk'))
+        data = {
+            "abstract": project_obj.description[:500],
+            "campus_affiliation": form_data.get('campus_affiliation', ''),
+            "directory_name": form_data.get('project_directory_name', ''),
+            "project_title": project_obj.title[:100],
+            "project_url": '',
+            "requested_size_tb": form_data.get('storage_space', ''),
+            "requester_email": self.request.user.email,
+            "requester_firstname": self.request.user.first_name,
+            "requester_lastname": self.request.user.last_name,
+            "start_date": start_date,
+            "submit_by": self.request.user.username,
+            "si": form_data.get('store_ephi', ''),
+            "service_type": "Slate-Project",
+            "account": form_data.get('account_number', ''),
+            "sub_account": '',
+            "fiscal_officer": '',
+            "faculty_advisor": ''
+        }
+        data = parse.urlencode(data)
+        try:
+            response = requests.post(
+                url=SLATE_PROJECT_MOU_SERVER,
+                headers={'Content-Type': 'application/x-www-form-urlencoded'},
+                data=data,
+                timeout=5
+            )
+            response.raise_for_status()
+        except requests.exceptions.Timeout:
+            logger.error(f'HTTP error: failed to send data to Slate Project MOU server: Request timed out')
+            form.add_error(None, 'Something went wrong processing your request. Please try again later')
+            return self.form_invalid(form)
+        except requests.HTTPError as http_error:
+            logger.error(f'HTTP error: failed to send data to Slate Project MOU server: {http_error}')
+            form.add_error(None, 'Something went wrong processing your request. Please try again later')
+            return self.form_invalid(form)
+
+        ldap_group = 'condo_' + form_data.get('project_directory_name', '')
+        form.cleaned_data['project_directory_name'] = '/N/project/' + form_data.get('project_directory_name', '')
+
+        http_response = super().form_valid(form)
+
+        ldap_group_type = AllocationAttributeType.objects.filter(name='LDAP Group')
+        if not ldap_group_type.exists():
+            logger.warning('LDAP Group allocation attribute type is missing')
+            return http_response
+
+        AllocationAttribute.objects.create(
+            allocation=self.allocation_obj,
+            allocation_attribute_type=ldap_group_type[0],
+            value=ldap_group)
+
+        return http_response
